+++ src/components/edit/thematic/RadiusSelect.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import i18n from '@dhis2/d2-i18n';
 import { Help } from '@dhis2/ui';
-import NumberField from '../../core/NumberField';
+import { NumberField } from '../../core';
 import { setRadiusLow, setRadiusHigh } from '../../../actions/layerEdit';
 import {
     THEMATIC_RADIUS_LOW,