+++ src/reducers/index.js
@@ -10,7 +10,6 @@ import dataSets from './dataSets';
 import dataTable from './dataTable';
 import dimensions from './dimensions';
 import download from './download.js';
-import earthEngine from './earthEngine';
 import layerEdit from './layerEdit';
 import indicators from './indicators';
 import indicatorGroups from './indicatorGroups';
@@ -35,6 +34,7 @@ import ui from './ui';
 import settings from './settings';
 import trackedEntityTypes from './trackedEntityTypes';
 import dataDownload from './dataDownload';
+import feature from './feature';
 
 export default combineReducers({
     alerts,
@@ -48,7 +48,6 @@ export default combineReducers({
     dataTable,
     dimensions,
     download,
-    earthEngine,
     indicators,
     indicatorGroups,
     interpretation,
@@ -73,4 +72,5 @@ export default combineReducers({
     settings,
     trackedEntityTypes,
     dataDownload,
+    feature,
 });