+++ src/util/favorites.js
@@ -2,6 +2,11 @@ import { isNil, omitBy, pick, isObject, omit } from 'lodash/fp';
 import { generateUid } from 'd2/uid';
 import { upgradeGisAppLayers } from './requests';
 import { getThematicLayerFromAnalyticalObject } from './analyticalObject';
+import {
+    EARTH_ENGINE_LAYER,
+    EXTERNAL_LAYER,
+    TRACKED_ENTITY_LAYER,
+} from '../constants/layers';
 
 // TODO: get latitude, longitude, zoom from map + basemap: 'none'
 const validMapProperties = [
@@ -110,6 +115,8 @@ const cleanLayerConfig = config => ({
 
 // TODO: This feels hacky, find better way to clean map configs before saving
 const models2objects = config => {
+    const { layer } = config;
+
     Object.keys(config).forEach(key => {
         config[key] = models.includes(key)
             ? pick(validModelProperties, config[key])
@@ -120,22 +127,32 @@ const models2objects = config => {
         config.rows = config.rows.map(cleanDimension);
     }
 
-    if (config.params) {
-        const { datasetId, params, filter, periodName } = config;
+    if (layer === EARTH_ENGINE_LAYER) {
+        const { datasetId: id, band, params, aggregationType, filter } = config;
 
-        // EE layer config
-        config.config = JSON.stringify({
-            id: datasetId,
+        const eeConfig = {
+            id,
             params,
-            image: filter ? filter[0].arguments[1] : null,
+            band,
+            aggregationType,
             filter,
-            periodName,
-        });
+        };
+
+        // Removes undefined keys before stringify
+        Object.keys(eeConfig).forEach(
+            key => eeConfig[key] === undefined && delete eeConfig[key]
+        );
+
+        config.config = JSON.stringify(eeConfig);
+
         delete config.datasetId;
         delete config.params;
         delete config.filter;
+        delete config.filters;
+        delete config.periodType;
         delete config.periodName;
-    } else if (config.layer === 'trackedEntity') {
+        delete config.aggregationType;
+    } else if (layer === TRACKED_ENTITY_LAYER) {
         config.config = JSON.stringify({
             relationships: config.relationshipType
                 ? {
@@ -191,7 +208,7 @@ export const cleanDimension = dim => ({
 const setExternalBasemap = config => {
     const { mapViews } = config;
     const externalBasemap = mapViews.find(view => {
-        if (view.layer === 'external') {
+        if (view.layer === EXTERNAL_LAYER) {
             if (typeof view.config === 'string') {
                 view.config = JSON.parse(view.config);
             }