+++ src/util/numbers.js
@@ -16,6 +16,9 @@ export const formatCount = count => {
 
 // Rounds a number to d decimals
 export const numberPrecision = d => {
+    if (d === undefined) {
+        return n => n;
+    }
     const m = Math.pow(10, d);
     return n => Math.round(n * m) / m;
 };