+++ webpack.config.js
@@ -61,6 +61,7 @@ const webpackConfig = {
                     /@dhis2\/ui-core/,
                     /@dhis2\/ui-forms/,
                     /@dhis2\/ui-widgets/,
+                    /@dhis2\/analytics/,
                 ],
                 loader: 'babel-loader',
                 query: {